+++ src/test/java/com/epam/reportportal/aspect/StepAspectFinishTest.java
@@ -35,9 +35,9 @@ import java.lang.reflect.Method;
 import java.util.UUID;
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.nullValue;
-import static org.mockito.Mockito.*;
+import static org.hamcrest.Matchers.*;
+import static org.mockito.Mockito.timeout;
+import static org.mockito.Mockito.verify;
 
 /**
  * @author <a href="mailto:vadzim_hushchanskou@epam.com">Vadzim Hushchanskou</a>
@@ -66,7 +66,7 @@ public class StepAspectFinishTest {
 	}
 
 	/*
-	 * Do not finish parent step inside nested step, leads to issue: https://github.com/reportportal/client-java/issues/97
+	 * Do not finish parent step inside nested step. This leads to issue: https://github.com/reportportal/client-java/issues/97
 	 */
 	@Test
 	public void verify_only_nested_step_finished_and_no_parent_steps() throws NoSuchMethodException {
@@ -75,11 +75,37 @@ public class StepAspectFinishTest {
 		aspect.finishNestedStep(method.getAnnotation(Step.class));
 		myLaunch.finish(TestUtils.standardLaunchFinishRequest());
 
+		ArgumentCaptor<String> finishIds = ArgumentCaptor.forClass(String.class);
 		ArgumentCaptor<FinishTestItemRQ> finishRQs = ArgumentCaptor.forClass(FinishTestItemRQ.class);
-		verify(client, timeout(1000).times(1)).finishTestItem(same(itemUuid), finishRQs.capture());
+		verify(client, timeout(1000).times(1)).finishTestItem(finishIds.capture(), finishRQs.capture());
+
+		String finishUuid = finishIds.getValue();
+		assertThat(finishUuid, sameInstance(itemUuid));
 
 		FinishTestItemRQ resultRq = finishRQs.getValue();
 		assertThat(resultRq.getStatus(), equalTo(ItemStatus.PASSED.name()));
+		assertThat(resultRq.getIssue(), nullValue());
+	}
+
+	/*
+	 * Do not finish parent step inside nested step. This leads to issue: https://github.com/reportportal/client-java/issues/97
+	 */
+	@Test
+	public void verify_only_nested_step_finished_and_no_parent_steps_on_step_failure() throws NoSuchMethodException {
+		Method method = StepAspectCommon.getMethod("testNestedStepSimple");
+		aspect.startNestedStep(StepAspectCommon.getJoinPoint(methodSignature, method), method.getAnnotation(Step.class));
+		aspect.failedNestedStep(method.getAnnotation(Step.class), new IllegalArgumentException());
+		myLaunch.finish(TestUtils.standardLaunchFinishRequest());
+
+		ArgumentCaptor<String> finishIds = ArgumentCaptor.forClass(String.class);
+		ArgumentCaptor<FinishTestItemRQ> finishRQs = ArgumentCaptor.forClass(FinishTestItemRQ.class);
+		verify(client, timeout(1000).times(1)).finishTestItem(finishIds.capture(), finishRQs.capture());
+
+		String finishUuid = finishIds.getValue();
+		assertThat(finishUuid, sameInstance(itemUuid));
+
+		FinishTestItemRQ resultRq = finishRQs.getValue();
+		assertThat(resultRq.getStatus(), equalTo(ItemStatus.FAILED.name()));
 		assertThat(resultRq.getIssue(), nullValue());
 	}
 }